requirements.txt
pandas
matplotlib
seaborn
fpdf
pyyaml
pytest